Add missing copy constructor
We have a copy assigned operator and without the corresponding copy
constructor we get a -Wdeprecated-copy warning with newer GCC versions.

Instead of adding the copy constructor we could have also dropped the copy
assignment since it is identical to what would be implied. It's not, however,
clear why it was added in the first place (it was added before we started
tracking development with git). Chances are, some old and broken compiler is
confused by the other assignment operator. So to be on the safe side we
rather add the copy constructor.